requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
asyncpg==0.23.0
attrs==21.2.0
cffi==1.14.5
chardet==4.0.0
click==8.0.1
discord-ext-menus @ git+https://github.com/Rapptz/discord-ext-menus@309c70222eec33acacb8c21bbb0fbe9bc57de661
discord.py==1.7.2
graphviz==0.16
humanize==3.5.0
idna==3.1
multidict==5.1.0
objgraph==3.5.0
pycparser==2.20
PyNaCl==1.4.0
pytz==2021.1
six==1.16.0
timeago==1.0.15
typing-extensions==3.10.0.0
wavelink==0.9.9
yarl==1.6.3
